added 6 additional func, modified casting, shorten a bunch of functions, fixed inconsistency etc etc

+++ libft/ft_strcat.c
@@ -6,7 +6,7 @@
 /*   By: trnguyen <trnguyen@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/03 11:10:04 by trnguyen          #+#    #+#             */
-/*   Updated: 2021/11/14 20:09:03 by trnguyen         ###   ########.fr       */
+/*   Updated: 2021/12/12 21:19:42 by trnguyen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strcat(char *s1, const char *s2)
 {
-	unsigned int	i;
-	unsigned int	j;
+	size_t	i;
+	size_t	j;
 
 	i = ft_strlen(s1);
 	j = 0;
