+++ libft/ft_strcpy.c
@@ -6,7 +6,7 @@
 /*   By: trnguyen <trnguyen@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 17:18:40 by trnguyen          #+#    #+#             */
-/*   Updated: 2021/11/14 20:09:15 by trnguyen         ###   ########.fr       */
+/*   Updated: 2021/12/12 21:31:19 by trnguyen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strcpy(char *dst, const char *src)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < ft_strlen(src))
